Add Github auth
I have decided to use Github as battlestations form of authentication
for the following reasons:
  - All developers should have an account
  - We can filter out anyone who is not part of a given organisation
  - We can then use teams inside of an organisation as roles for a user

When Github returns us a code through the redirect url we use it to
generate a User struct being making calls to the Github API. We retrieve
the user's details, whether they are a member of the target organisation
and if they are their list of teams. If they are not a member of the
target organsiation then an error is raised and the attempt to build a
User struct is aborted.

In order to retrieve the list of teams a user is a part of I have had to
drop down to making the http request manually, this is because I could
not find a function in the go-github lib that would do this form me. It
encodes to the same go-github Team type so all should be lovely.