Integrate the api and Github auth
I have added the routes required to do the Github auth dance. It is
using the gorilla/sessions session store to hold the User structs for
each session. This is currently using the file store but should probably
move to Redis or something later.

I have created a wrapping function to retrieve the User struct from the
session store and if it exists pass through, otherwise redirect to
/authorize to start authentication.

The /authorized endpoint attempts to get a User struct, if it succeeds
then it will redirect to / after storing the struct in a session var, if
not then it will print out the error message.

Battlestations now requires two keys to auth and encrypt the cookies
used by the session store. The path to these keys should be passed
through in the config JSON.